+++ db/upgrade.php
@@ -246,5 +246,20 @@ function xmldb_enrol_autoenrol_upgrade($oldversion) {
         upgrade_plugin_savepoint(true, 2021101800, 'enrol', 'autoenrol');
     }
 
+    if ($oldversion < 2021112600) {
+        $instances = $DB->get_records_select('enrol', 'enrol = \'autoenrol\' AND enrolperiod > 0');
+        if (!empty($instances)) {
+            foreach ($instances as $instance) {
+                if ($enrolments = $DB->get_records('user_enrolments', array('enrolid' => $instance->id))) {
+                    foreach ($enrolments as $enrolment) {
+                        $enrolment->timeend = $enrolment->timestart + $instance->enrolperiod;
+                        $DB->update_record('user_enrolments', $enrolment);
+                    }
+                }
+            }
+        }
+        upgrade_plugin_savepoint(true, 2021112600, 'enrol', 'autoenrol');
+    }
+
     return true;
 }