+++ lib.php
@@ -275,7 +275,13 @@ class enrol_autoenrol_plugin extends enrol_plugin {
             throw new coding_exception('Invalid enrol instance type!');
         }
         if ($this->enrol_allowed($user, $instance)) {
-            $this->enrol_user($instance, $user->id, $instance->roleid, time(), 0);
+            $timestart = time();
+            if ($instance->enrolperiod) {
+                $timeend = $timestart + $instance->enrolperiod;
+            } else {
+                $timeend = 0;
+            }
+            $this->enrol_user($instance, $user->id, $instance->roleid, $timestart, $timeend);
             $this->process_group($instance, $user);
             // Send welcome message.
             if ($instance->customint7 != ENROL_DO_NOT_SEND_EMAIL) {